Removed an assert

+++ as/cNFT/assembly/royalty_payout.ts
@@ -26,7 +26,7 @@ export function nft_payout(
     let royalty_sb_keys = token_royalty.split_between.keys()
     let royalty_sb_size = token_royalty.split_between.size
 
-    assert(royalty_sb_size <= max_len_payout, "Royalty size greater than max len")
+    assert(u32(royalty_sb_size) <= max_len_payout, "Royalty size greater than max len")
 
     /** Go through each key and value in the royalty object */
     for (let i = 0; i < royalty_sb_size; i++) {
